Adding needed reference to `Statiq.Markdown`
The `RenderMarkdown()` method is found in the `Statiq.Markdown` nuget package.  I was using these docs to see how to add a markdown renderer; and this is the part I got stuck on because I couldn't find the `RenderMarkdown` class in the `Statiq.App` namespace; turns out, it's in another package entirely.  Adding this using statement so that the code would compile (I hope?).